requirements.txt
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.0.4
django-appconf==1.0.4
django-crispy-forms==1.9.0
django-filter==2.2.0
django-heroku==0.3.1
django-import-export==2.0.2
django-mediumeditor==1.0.0
et-xmlfile==1.0.1
facebook-sdk==3.1.0
firebase==3.0.1
gunicorn==20.0.4
idna==2.9
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
short-url==1.2.2
sqlparse==0.3.1
tablib==1.1.0
urllib3==1.25.8
whitenoise==5.0.1
wincertstore==0.2
xlrd==1.2.0
xlwt==1.3.0
